Rename file, address code-review comments, add docstrings
